1. Solved the problem of dose calculation stall. It was because the epsilon was too small, causing some unexpected behavior. Changing the epsilong from 1e-4 to 1e-3 solves the problem. 2. Benchmarked the dose calculation step, it took 73 seconds to calculate the 400+ beams. Found that the memory allocation et cl could be largely improved. To implement it later.

+++ IMRTOpt/include/IMRTDoseMat.cuh
@@ -50,7 +50,7 @@ namespace IMRT {
             }
         }
 
-        bool dense2sparse(float* d_dense, int num_rows, int num_cols, int ld);
+        bool dense2sparse(float* d_dense, size_t num_rows, size_t num_cols, size_t ld);
 
         cusparseSpMatDescr_t matA;
         int* d_csr_offsets;
